+++ lib/main/atom/commands/goToDeclaration.tsx
@@ -5,11 +5,14 @@ import * as etch from "etch"
 
 const prevCursorPositions: FileLocationQuery[] = []
 
-function open(item: {file: string; start: {line: number; offset: number}}) {
-  atom.workspace.open(item.file, {
+async function open(item: {file: string; start: {line: number; offset: number}}) {
+  const editor = await atom.workspace.open(item.file, {
     initialLine: item.start.line - 1,
     initialColumn: item.start.offset - 1,
   })
+  if (atom.workspace.isTextEditor(editor)) {
+    editor.scrollToCursorPosition({center: true})
+  }
 }
 
 commands["atom-text-editor"]["typescript:go-to-declaration"] = deps => ({