+++ CHANGELOG.md
@@ -1,5 +1,16 @@
+## 12.2.0
+
+-   Merge pull request #1402 "Center cursor after open in goToDeclaration" by Kai Curtis
+-   Merge pull request #1280 "Re-adding semantic view" by russaa
+
+## 12.1.2
+
+-   Use BufferedNodeProcess to spawn tsc (fixes 'init tsconfig' on
+    Windows)
+
 ## 12.1.1
-* Fix occurence markers
+
+-   Fix occurrence markers
 
 ## 12.1.0